Reorganiza los proyectos, añadiendo un proyecto maven llamado entidads y cambiando de nombre al prollecto con los DTOs
