+++ .gitignore
@@ -10,3 +10,5 @@
 /PersistenciaListasBazar/nbproject/
 >>>>>>> Stashed changes
 /PersistenciaBazar/target/
+/ObjetosDTO/target/
+/Entidades/target/
