+++ Entidades/src/main/java/entidades/Compra.java
