+++ Entidades/src/main/java/entidades/DetalleCompra.java
