+++ Entidades/src/main/java/entidades/DetalleVenta.java
@@ -1,5 +1,5 @@
 /*
- * InventarioProducto.java
+ * DetalleVenta.java
  */
 package entidades;
 
@@ -10,17 +10,17 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
-import javax.persistence.OneToOne;
+import javax.persistence.ManyToOne;
 import javax.persistence.Table;
-import objetosNegocio.InventarioProductoDTO;
+import objetosDTO.DetalleVentaDTO;
 
 /**
  * 
  * @author Juventino López García
  */
 @Entity
-@Table(name = "inventario_productos")
-public class InventarioProducto implements Serializable {
+@Table(name = "detalle_ventas")
+public class DetalleVenta implements Serializable {
 
     private static final long serialVersionUID = 1L;
     @Id
@@ -30,11 +30,15 @@ public class InventarioProducto implements Serializable {
     @Column(name = "cantidad")
     private Integer cantidad;
     
-    @OneToOne
-    @JoinColumn(name = "id_producto")
+    @ManyToOne
+    @JoinColumn(name = "producto")
     private Producto producto;
+    
+    @Column(name = "precio_producto")
+    private Float precioProducto;
 
-    public InventarioProducto() {
+    public DetalleVenta() {
+        
     }
     
     public Long getId() {
@@ -61,12 +65,21 @@ public class InventarioProducto implements Serializable {
         this.producto = producto;
     }
     
-    public InventarioProductoDTO toDTO() {
-        InventarioProductoDTO i = new InventarioProductoDTO();
-        i.setCantidad(cantidad);
-        i.setProducto(producto.toDTO());
+    public Float getPrecioProducto() {
+        return this.precioProducto;
+    }
+    
+    public void setPrecioProducto(Float precio) {
+        this.precioProducto = precio;
+    }
+    
+    public DetalleVentaDTO toDTO() {
+        DetalleVentaDTO d = new DetalleVentaDTO();
+        d.setCantidad(cantidad);
+        d.setProducto(producto.toDTO());
+        d.setPrecioProducto(this.precioProducto);
         
-        return i;
+        return d;
     }
 
 }
