+++ Entidades/src/main/java/entidades/Producto.java
@@ -9,7 +9,7 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import objetosNegocio.ProductoDTO;
+import objetosDTO.ProductoDTO;
 
 /**
  *
