+++ ObjetosDTO/src/main/java/objetosDTO/ProductoDTO.java
@@ -1,7 +1,7 @@
 /*
  * ProductoDTO.java
  */
-package objetosNegocio;
+package objetosDTO;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
