+++ ObjetosDTO/src/main/java/objetosDTO/UsuarioDTO.java
@@ -1,7 +1,7 @@
 /*
  * UsuarioDTO.java
  */
-package objetosNegocio;
+package objetosDTO;
 
 import java.time.LocalDateTime;
