+++ ObjetosDTO/src/main/java/objetosDTO/VentaDTO.java
@@ -1,7 +1,7 @@
 /*
  * VentaDTO.java
  */
-package objetosNegocio;
+package objetosDTO;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
