+++ PersistenciaBazar/src/main/java/dao/GestorUsuarios.java
@@ -10,7 +10,7 @@ import java.util.regex.Pattern;
 import javax.persistence.EntityManager;
 import javax.persistence.NoResultException;
 import javax.persistence.TypedQuery;
-import objetosNegocio.UsuarioDTO;
+import objetosDTO.UsuarioDTO;
 import subsistemas.excepciones.DAOException;
 import subsistemas.interfaces.IGestorUsuarios;
